backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-multipart==0.0.20
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-nmap==0.7.1
aiofiles==24.1.0
email-validator==2.2.0
